requirements.txt
asgiref==3.4.1
async-timeout==4.0.2
Deprecated==1.2.13
Django==3.2.14
django-redis==5.2.0
django-restframework==0.0.1
djangorestframework==3.13.1
pyparsing==3.0.9
pytz==2022.1
redis==4.3.4
sqlparse==0.4.2
typing_extensions==4.1.1
wrapt==1.14.1
zipp==3.6.0
flake8